Update package exports to explicit
