+++ packages/forms/src/index.ts
@@ -1,8 +1,9 @@
 // Components
-export * from './components/Form';
-export * from './components/FormField';
-export * from './components/FormSubmit';
-export * from './components/TextInput';
+export { Form } from './components/Form';
+export { FormField } from './components/FormField';
+export { FormSubmit } from './components/FormSubmit';
+export { TextInput } from './components/TextInput';
 
 // Hooks
-export * from './hooks/useFormField';
+export { useFormField } from './hooks/useFormField';
+export { useTextInput } from './hooks/useTextInput';
