--- packages/internal/src/checks/index.ts
@@ -1,8 +1,0 @@
-export { checkAccessibilityRole } from './checkAccessibilityRole';
-export { checkFocusTrap } from './checkFocusTrap';
-export { checkForAccessibilityState } from './checkForAccessibilityState';
-export { checkMinimumSize } from './checkMinimumSize';
-export { contrastChecker } from './contrastChecker';
-export { noUndefinedProperty } from './noUndefinedProperty';
-export { uppercaseChecker } from './uppercaseChecker';
-export { uppercaseStringChecker } from './uppercaseStringChecker';
